Move addToChangeLog to common database methods
